+++ compiler/lib/link_js.mli
@@ -20,6 +20,8 @@
 val link :
      output:out_channel
   -> linkall:bool
+  -> mklib:bool
+  -> toplevel:bool
   -> files:string list
   -> resolve_sourcemap_url:bool
   -> source_map:(string option * Source_map.t) option